Introduce SMART map abstraction
The concept of a map is to allow the protocol specific layers to point
to each attribute in memory and then describe it in terms of endianess,
byte count, and whether it has associated threshold values. This will
allow the print function to be the same for all protocols despite the
attribute size and location differences.