Generalize device_read into device_read_log
Instead of assuming that device_read() returns the protocol specific data
associated with the SMART log page(s), change the API to device_read_log()
which includes the specific log page id as a parameter. This allows
moving the logic to probe SCSI log pages into OS-generic code as well as
the possibility of pulling vendor specific log pages in the future.