Fix ATA attribute max
The data returned by SMART READ DATA uses bytes 0-361 for vendor
specific data. This is where attributes can be found. Skipping the two
bytes of version and knowing that each entry is 12 bytes, there are a
maximum of 30 entries.

The application was assuming the entire allocated buffer contained
attributes and mistakenly interpreted other fields of the data
structure as attributes. This lead to reports of 're-used attributes'.

Fix is to return 30 from __smart_attr_max_ata().

resolved bug #4