Document and fix ATA threshold option
The --threshold option prints the 4 bytes between the attribute ID and
the raw attribute value. Document in the README these make up the flags
(2 bytes), current value, and worst value.

The existing code had two usability issues. First, it printed the 4 byte
values individually instead of combining the first two bytes into the
flags value. Fix is to group the bytes into the natural 3 values.
Second, the application always printed the raw value last, but this
means its field number changes depending on the presence of the
threshold values. Fix is to print the threshold values last.

resolved bug #6