Fix the maximum attribute calculations
Allocation of the smart map uses the attr_count variable. But, in the
case of NVMe and SCSI, the code was returning zero opening the
possibility of a buffer overrun as the subsequent code didn't check for
this case.

Fix is to return a non-zero default value for NVMe and SCSI maximum
attribute calculations in addition to checking for a zero count in the
map allocation code itself.