Add ATA SMART Return Status
This adds support for the ATA SMART Return Status command to retrieve
the reliability status of the device. The command encodes the status in
the LBA high and mid registers, and either indicates:
 - The device has detected a threshold exceeded condition (a.k.a. bad)
 - The captive self-test completed without error (a.k.a. good)

While the command does not return data, the program encodes the above
status into the data buffer at offset / attribute 0x0. A good status
(i.e. LBA high/mid = 0xc24f) is encoded as a 0. A bad status (i.e. LBA
high/mid = 0x2cf4) is encoded as a 1.

Some applications use this command as a sort of top-level SMART health
indication.

addresses issue #5