feat: make attribute decode an option
Instead of unconditionally decoding the attribute ID's, add two new
options:
  --decode, -d    : decodes the attribute ID into text
  --no-decode, -D : don't decode the attribute ID into text

If the program is invoked as 'smart', it defaults to the original
behavior of only printing values (i.e. '--no-decode'). But if the
program is invoked as 'diskhealth', it defaults to decoding the
attribute ID and printing the value (i.e. '--decode'). Note that the
'decode' behavior skips printing the page ID and attribute ID.

While in the neighborhood, create macros for the flag values passed to
smart_open().